@@ -1,8 +1,8 @@
-﻿using Newtonsoft.Json;
 using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
 
 namespace GrpcJsonTranscoder.Internal.Http
 {
@@ -23,13 +23,16 @@ namespace GrpcJsonTranscoder.Internal.Http
         protected override async Task SerializeToStreamAsync(Stream stream, TransportContext context)
         {
             var writer = new StreamWriter(stream);
+
             await writer.WriteAsync(_result);
+
             await writer.FlushAsync();
         }
 
         protected override bool TryComputeLength(out long length)
         {
             length = _result.Length;
+
             return true;
         }
     }
