@@ -1,8 +1,9 @@
-﻿namespace GrpcJsonTranscoder.Internal
+namespace GrpcJsonTranscoder.Internal
 {
     public class NameAndValue
     {
         public string Name { get; set; }
+
         public string Value { get; set; }
     }
 }
