@@ -11,14 +11,12 @@ namespace GrpcJsonTranscoder
         {
             //AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
 
-            using (var scope = services.BuildServiceProvider().CreateScope())
-            {
-                var svcProvider = scope.ServiceProvider;
-                var config = svcProvider.GetRequiredService<IConfiguration>();
-                services.Configure<GrpcMapperOptions>(config.GetSection("GrpcJsonTranscoder"));
-                services.AddSingleton(resolver => addGrpcAssembly.Invoke());
-                return services;
-            }
+            using var scope = services.BuildServiceProvider().CreateScope();
+            var svcProvider = scope.ServiceProvider;
+            var config = svcProvider.GetRequiredService<IConfiguration>();
+            services.Configure<GrpcMapperOptions>(config.GetSection("GrpcJsonTranscoder"));
+            services.AddSingleton(resolver => addGrpcAssembly.Invoke());
+            return services;
         }
     }
 }
